+++ packages/contracts-nest/src/grpc.helpers.ts
@@ -16,7 +16,18 @@ export enum GRPC_SERVICES {
   EVENT_QUERY_SERVICE = 'EventQueryService',
   TAG_COMMAND_SERVICE = 'TagCommandService',
   TAG_QUERY_SERVICE = 'TagQueryService',
-  SOCIAL_SERVICE = 'SocialService',
+  SOCIAL_USER_NODE_COMMAND_SERVICE = 'SocialUserNodeCommandService',
+  SOCIAL_USER_NODE_QUERY_SERVICE = 'SocialUserNodeQueryService',
+  RELATIONSHIP_COMMAND_SERVICE = 'RelationshipCommandService',
+  RELATIONSHIP_QUERY_SERVICE = 'RelationshipQueryService',
+  PUBLICATION_COMMAND_SERVICE = 'PublicationCommandService',
+  PUBLICATION_QUERY_SERVICE = 'PublicationQueryService',
+  INTERACTION_COMMAND_SERVICE = 'InteractionCommandService',
+  INTERACTION_QUERY_SERVICE = 'InteractionQueryService',
+  PARTICIPATION_COMMAND_SERVICE = 'ParticipationCommandService',
+  PARTICIPATION_QUERY_SERVICE = 'ParticipationQueryService',
+  EVENT_POST_LINK_COMMAND_SERVICE = 'EventPostLinkCommandService',
+  EVENT_POST_LINK_QUERY_SERVICE = 'EventPostLinkQueryService',
 }
 
 export enum GRPC_MICROSERVICES {